test fault tolerance: size_t data type for IDs
(cherry picked from commit 1ab8477aed3b816a0fb28c4cdb2762c43b186d8b)

+++ distributedcombigrid/tests/test_ftolerance.cpp
@@ -305,7 +305,7 @@ for (size_t i = 1; i < ncombi; ++i){
         std::cout << "failed group detected at combi iteration " << i-1<< std::endl;
         //manager.recover();
 
-        std::vector<int> faultsID;
+        std::vector<size_t> faultsID;
 
         //vector with pointers to managers of failed groups
         std::vector< ProcessGroupManagerID> groupFaults;
@@ -313,7 +313,7 @@ for (size_t i = 1; i < ncombi; ++i){
 
         /* call optimization code to find new coefficients */
         const std::string prob_name = "interpolation based optimization";
-        std::vector<int> redistributeFaultsID, recomputeFaultsID;
+        std::vector<size_t> redistributeFaultsID, recomputeFaultsID;
         manager.recomputeOptimumCoefficients(prob_name, faultsID,
                                              redistributeFaultsID, recomputeFaultsID);
         for ( auto id : redistributeFaultsID ) {
